requirements.txt
azure-common==1.1.28
azure-core==1.30.2
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
Django==4.2.5
djangorestframework==3.14.0
psycopg2==2.9.9
whitenoise==6.6.0
openai==1.30.1
django-cors-headers==4.3.1
pip-system-certs==4.0
drf-nested-routers==0.94.1
python-dotenv==1.0.1
llama-index==0.10.37
llama-index-lms-azure-openai==0.1.8
llama-index-embeddings-azure-openai==0.1.9
llama-index-experimental==0.1.3
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
pytz==2023.3.post1
